fix: rename package errors => error
