+++ kafkamgmt/README.md
@@ -74,7 +74,7 @@ kafkas, _, err := client.DefaultApi.GetKafkas(ctx).Execute()
 Checking specific error codes
 
 ```go
-import "github.com/redhat-developer/app-services-sdk-go/registrymgmt/apiv1/errors"
+import "github.com/redhat-developer/app-services-sdk-go/registrymgmt/apiv1/error"
 
 if errors.IsAPIError(err, kafkamgmt.ERROR_4){
  // Do something
@@ -84,7 +84,7 @@ if errors.IsAPIError(err, kafkamgmt.ERROR_4){
 Obtaining API error 
 
 ```go
-import "github.com/redhat-developer/app-services-sdk-go/registrymgmt/apiv1/errors"
+import "github.com/redhat-developer/app-services-sdk-go/registrymgmt/apiv1/error"
 
 apiError := errors.GetAPIError(err)
 fmt.Println(apiError.Code)
