+++ kafkamgmt/apiv1/error/errors.go
@@ -1,5 +1,5 @@
 
-package errors
+package error
 
 // kafkamgmt error codes 
 type ErrorCode string
