+++ registrymgmt/apiv1/error/error_sdk.go
@@ -1,19 +1,19 @@
-package errors
+package error
 
 import (
 	"errors"
 
-	connectormgmtclient "github.com/redhat-developer/app-services-sdk-go/connectormgmt/apiv1/client"
+	registrymgmtclient "github.com/redhat-developer/app-services-sdk-go/connectormgmt/apiv1/client"
 )
 
 // GetAPIError gets a strongly typed error from an error
-func GetAPIError(err error) *connectormgmtclient.Error {
-	var openapiError connectormgmtclient.GenericOpenAPIError
+func GetAPIError(err error) *registrymgmtclient.Error {
+	var openapiError registrymgmtclient.GenericOpenAPIError
 
 	if ok := errors.As(err, &openapiError); ok {
 		errModel := openapiError.Model()
 
-		transformedError, ok := errModel.(connectormgmtclient.Error)
+		transformedError, ok := errModel.(registrymgmtclient.Error)
 		if !ok {
 			return nil
 		}
@@ -24,7 +24,6 @@ func GetAPIError(err error) *connectormgmtclient.Error {
 }
 
 // IsAPIError returns true if the error contains the errCode
-// Error code is an code that is returned by the API
 func IsAPIError(err error, code string) bool {
 	mappedErr := GetAPIError(err)
 	if mappedErr == nil {
